more recoil tidy
