+++ src/recoil/selectors.ts
@@ -8,26 +8,11 @@ import {
   PlatformInfoJSON,
   RequiredFileInfo,
   Screenshot,
-  VideoJSON,
 } from "../types"
 import { renderBinImage } from "../components/utils/renderBinImage"
 import { fileSystemInvalidationAtom } from "./atoms"
 import { getVersion } from "@tauri-apps/api/app"
 import { decodeDataParams } from "../components/utils/decodeDataParams"
-
-export const VideoJSONSelectorFamily = selectorFamily<VideoJSON, string>({
-  key: "VideoJSONSelectorFamily",
-  get:
-    (coreName) =>
-    async ({ get }) => {
-      get(fileSystemInvalidationAtom)
-      const jsonText = await invoke<string>("read_text_file", {
-        path: `Cores/${coreName}/video.json`,
-      })
-
-      return JSON.parse(jsonText) as VideoJSON
-    },
-})
 
 export const DataJSONSelectorFamily = selectorFamily<DataJSON, string>({
   key: "DataJSONSelectorFamily",
@@ -78,7 +63,7 @@ export const RequiredFileInfoSelectorFamily = selectorFamily<
               path: `${path}/${filename}`,
             }),
             type: "core",
-          }
+          } as RequiredFileInfo
         })
       )
 
@@ -106,8 +91,6 @@ export const RequiredFileInfoSelectorFamily = selectorFamily<
               extension: ".json",
             })
 
-            console.log({ files })
-
             return await Promise.all(
               files.map(async (f) => {
                 const response = await invoke<string>("read_text_file", {
@@ -116,8 +99,6 @@ export const RequiredFileInfoSelectorFamily = selectorFamily<
 
                 const instanceFile = JSON.parse(response) as InstanceDataJSON
                 const dataPath = instanceFile.instance.data_path
-
-                console.log({ instanceFile })
 
                 return await Promise.all(
                   instanceFile.instance.data_slots.map(
@@ -139,7 +120,7 @@ export const RequiredFileInfoSelectorFamily = selectorFamily<
                           path: fullPath,
                         }),
                         type: "instance",
-                      }
+                      } as RequiredFileInfo
                     }
                   )
                 )
@@ -149,66 +130,6 @@ export const RequiredFileInfoSelectorFamily = selectorFamily<
       )
 
       return [...fileInfo, ...instanceFileInfo.flat(3)]
-    },
-})
-
-export const screenshotsListSelector = selector<string[]>({
-  key: "screenshotsListSelector",
-  get: async ({ get }) => {
-    get(fileSystemInvalidationAtom)
-    return await invoke<string[]>("list_files", {
-      path: "Memories/Screenshots",
-    })
-  },
-})
-
-export const SingleScreenshotSelectorFamily = selectorFamily<
-  Screenshot | null,
-  string
->({
-  key: "SingleScreenshotSelectorFamily",
-  get:
-    (fileName) =>
-    async ({ get }) => {
-      get(fileSystemInvalidationAtom)
-
-      const data = await invoke<number[]>("read_binary_file", {
-        path: `Memories/Screenshots/${fileName}`,
-      })
-      const buf = new Uint8Array(data)
-      const file = new File([buf], fileName, { type: "image/png" })
-      const metadataBuffer = buf.slice(buf.length - 528)
-
-      let utf8decoder = new TextDecoder()
-      // The unpacking here might not be right if there's unused ranges
-
-      let authorName = utf8decoder
-        .decode(metadataBuffer.slice(0, 16 * 2))
-        .replaceAll("\u0000", "")
-
-      let coreName = utf8decoder
-        .decode(metadataBuffer.slice(16 * 2, 16 * 4))
-        .trim()
-        .replaceAll("\u0000", "")
-
-      let gameName = utf8decoder
-        .decode(metadataBuffer.slice(16 * 6, 16 * 20))
-        .trim()
-        .replaceAll("\u0000", "")
-
-      let platformName = utf8decoder
-        .decode(metadataBuffer.slice(metadataBuffer.length - 16 * 10))
-        .replaceAll("\u0000", "")
-
-      return {
-        file_name: fileName,
-        file,
-        game: gameName,
-        platform: platformName,
-        timestamp: new Date(file.lastModified),
-        author: authorName,
-        core: coreName,
-      }
     },
 })
